fix(economy,zhtp): restore u128::MAX bounds; fix json! panic in API handlers
Review follow-ups on the original commit:

- edge_case_tests: two tests were capped at u64::MAX on the grounds that
  u128::MAX trips the unchecked multiply in calculate_fee. That was wrong
  for these two — neither reaches calculate_fee (Transaction::new_payment
  routes through the saturating calculate_dao_fee). They pass at true
  u128::MAX, and capping them lost the regression coverage that would
  catch calculate_dao_fee's saturating_mul being reverted. Restored.

- zhtp API: get_stats() and get_economic_status() feed max_supply
  (u128::MAX by default) straight into json!, so both panicked with
  'number out of range' on every call — the same bug fixed in
  get_economic_stats(). Emit u128 as decimal strings there too. No
  consumer can have depended on the number shape: the handlers panicked
  unconditionally.

The caps in fee_processing.rs remain — calculate_dao_fee_distribution
genuinely uses an unchecked multiply, so u64::MAX is the honest ceiling
there until #2875 lands.

+++ lib-economy/tests/edge_case_tests.rs
@@ -27,7 +27,7 @@ mod edge_case_tests {
     #[test]
     fn test_maximum_amount_transactions() {
         // Test transaction with maximum possible amount
-        let max_amount = u64::MAX as u128;
+        let max_amount = u128::MAX;
         let tx =
             Transaction::new_payment([1u8; 32], [2u8; 32], max_amount, Priority::Normal).unwrap();
 
@@ -187,14 +187,14 @@ mod edge_case_tests {
         assert!(!wallet.can_afford(1));
 
         // Test with maximum balance
-        wallet.available_balance = u64::MAX as u128;
-        assert!(wallet.can_afford(u64::MAX as u128));
+        wallet.available_balance = u128::MAX;
+        assert!(wallet.can_afford(u128::MAX));
         assert!(wallet.can_afford(1)); // can_afford is a plain >= check
 
         // Test total balance calculation with maximum values
-        wallet.staked_balance = u64::MAX as u128 / 3;
-        wallet.pending_rewards = u64::MAX as u128 / 3;
-        wallet.available_balance = u64::MAX as u128 / 3;
+        wallet.staked_balance = u128::MAX / 3;
+        wallet.pending_rewards = u128::MAX / 3;
+        wallet.available_balance = u128::MAX / 3;
 
         let total = wallet.total_balance();
         assert!(total >= wallet.available_balance); // Should not underflow
